% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/livechatCreateAccount.R
\name{livechatCreateAccount}
\alias{livechatCreateAccount}
\title{Setup connection with LiveChat's REST API using email and an api key}
\usage{
livechatCreateAccount(email, api_key)
}
\arguments{
\item{email}{Email address used to login from \href{https://my.livechatinc.com/agents/api-key}{Live Chat API Docs}}

\item{api_key}{Your API Key from \href{https://my.livechatinc.com/agents/api-key}{Live Chat API Docs}}
}
\value{
A list of 2, the first item is called email and the second item is called api_key
}
\description{
Setup connection with LiveChat's REST API using email and an api key
}
\examples{
# account <- livechatCreateAccount("email_here", "api_key_here")
# account$email
# account$api_key
}

