% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/livechatGetAgents.R
\name{livechatGetAgents}
\alias{livechatGetAgents}
\title{Create a list of agents with LiveChat credentials as parameters}
\usage{
livechatGetAgents(account)
}
\arguments{
\item{account}{defined by the livechatCreateAccount function}
}
\value{
A date frame of agents
}
\description{
Create a list of agents with LiveChat credentials as parameters
}
\examples{
# account <- livechatCreateAccount("email_here", "api_key_here")
# livechatGetAgents(account)
}

