% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/livechatGetCannedResponses.R
\name{livechatGetCannedResponses}
\alias{livechatGetCannedResponses}
\title{Create a list of canned responses with LiveChat credentials as parameters}
\usage{
livechatGetCannedResponses(account, group = 0)
}
\arguments{
\item{account}{defined by the livechatCreateAccount function}

\item{group}{defaults to 0, see \href{https://developers.livechatinc.com/rest-api/#!canned-responses}{LiveChat REST API}}
}
\value{
a data frame of canned responses
}
\description{
Create a list of canned responses with LiveChat credentials as parameters
}
\examples{
# account <- livechatCreateAccount("email_here", "api_key_here")
# livechatGetCannedResponses(account, group = 0)
}

