% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/livechatGetChats.R
\name{livechatGetChats}
\alias{livechatGetChats}
\title{Create a list that contains 4 data frames containing chat sessions and raw chat text.}
\usage{
livechatGetChats(account, date_from)
}
\arguments{
\item{account}{defined by the livechatCreateAccount function}

\item{date_from}{pull all chats from the date in the format YYYY-MM-DD}
}
\value{
A list of 4. The being a data.table of chat sessions, messages, postchat surveys and events
}
\description{
Create a list that contains 4 data frames containing chat sessions and raw chat text.
}
\examples{
# account <- livechatCreateAccount("email_here", "api_key_here")
# livechat_data <- livechatGetChats(account, date_from = "2016-02-23")
# str(livechat_data)
# livechat_data[[1]] \%>\% str # chat_sessions
# livechat_data[[2]] \%>\% str # messages
# livechat_data[[3]] \%>\% str # postchat_survey
# livechat_data[[4]] \%>\% str # events
}

