% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/livechatGetGoals.R
\name{livechatGetGoals}
\alias{livechatGetGoals}
\title{Create a list of goals with LiveChat credentials as parameters}
\usage{
livechatGetGoals(account)
}
\arguments{
\item{account}{defined by the livechatCreateAccount function}
}
\value{
a data frame of goals
}
\description{
Create a list of goals with LiveChat credentials as parameters
}
\examples{
# account <- livechatCreateAccount("email_here", "api_key_here")
# livechatGetGoals(account)
}

