% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/livechatGetGreetings.R
\name{livechatGetGreetings}
\alias{livechatGetGreetings}
\title{Create a list of greetings with LiveChat credentials as parameters}
\usage{
livechatGetGreetings(account, group = 0)
}
\arguments{
\item{account}{defined by the livechatCreateAccount function}

\item{group}{defaults to 0, see \href{https://developers.livechatinc.com/rest-api/#!get-greetings}{LiveChat REST API}}
}
\value{
a data frame of greetings
}
\description{
Create a list of greetings with LiveChat credentials as parameters
}
\examples{
# account <- livechatCreateAccount("email_here", "api_key_here")
# livechatGetGreetings(account, group = 0)
}

